Rename syntax file
