+++ tests/test_check_function_def.py
@@ -2,7 +2,7 @@ import pytest
 import helper
 
 from pythonwhat.local import setup_state
-from pythonwhat.check_syntax import v2_check_functions
+from pythonwhat.sct_syntax import v2_check_functions
 globals().update(v2_check_functions)
 
 @pytest.mark.parametrize('stu, passes', [
