+++ tests/test_check_list_comp.py
@@ -1,7 +1,7 @@
 import helper
 import pytest
 from pythonwhat.local import setup_state
-from pythonwhat.check_syntax import v2_check_functions
+from pythonwhat.sct_syntax import v2_check_functions
 globals().update(v2_check_functions)
 
 @pytest.mark.parametrize('stu, passes', [
